+++ packages/client/src/api/deployment-lock.api.ts
@@ -11,7 +11,7 @@ export const getIsLocked = async ({
   deploymentId,
 }: GetIsLockedOptions): Promise<boolean> => {
   const { data } = await client
-    .get<unknown, { data: string }>("deployment-locks/is-locked", {
+    .get<string>("deployment-locks/is-locked", {
       params: { deploymentId },
     })
     .catch((error) => {