+++ packages/custom-checks/src/get-snapshots-from-test-run.ts
@@ -55,7 +55,7 @@ export const getSnapshotsFromTestRun = async ({
     params.append("snapshotTypes", snapshotType);
   }
   const { data } = await client
-    .get<unknown, { data: SnapshotDownloadUrlsResponse }>(
+    .get<SnapshotDownloadUrlsResponse>(
       `test-runs/${testRunId}/custom-check-snapshots-download-urls?${params.toString()}`,
     )
     .catch((error) => {